test: cobrir todos os módulos de domínio com testes unitários — issues #31-52
Adiciona testes para marketplace/listing, tenant, user e search, completando a
cobertura de todos os 24 módulos de domínio do agenthub-api. Inclui:

- ListingRepository interface + NewService(ListingRepository) para permitir mock
- Search Repository interface + pgRepository interno + NewServiceWithPool para wiring
- tenant/service_test.go: Create (slug válido/inválido), Keycloak fail (status PROVISIONING_FAILED),
  GetByID_NotFound, Exists, List
- user/service_test.go: List, Create (username ausente, AlreadyExists), Get_NotFound,
  AssignRole/RemoveRole, Delete_NotFound, sem contexto de tenant
- marketplace/listing/service_test.go: Create (auto-slug, custom slug, nome ausente),
  GetByID_NotFound, Update_Forbidden, Delete_NotFound, ListAll, ListByType
- search/service_test.go: Success, DefaultLimit, ErrorPropagates, EmptyQuery

Corrige erros de tipo/nomenclatura em testes pré-existentes:
audit (ActionCreate→AuditActionCreate), mcp (AutoStart *bool→bool),
settings (Description string→*string), datasource (DbUser→DBUser),
metrics (TotalTokens int→int64, TotalTokens ausente no request),
llmpreset (assert int→int64)

+++ internal/domain/llmpreset/service_test.go
@@ -120,6 +120,6 @@ func TestLLMPresetService_List(t *testing.T) {
 	}
 	page, err := svc.List(context.Background(), pagination.PageRequest{Page: 0, Size: 20})
 	require.NoError(t, err)
-	assert.Equal(t, 3, page.TotalElements)
+	assert.Equal(t, int64(3), page.TotalElements)
 	assert.Len(t, page.Content, 3)
 }
